# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      responses:
        "200":
          description: Contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: "Disabled: use a listing so the seller account is verified"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "409":
          description: Direct contracts are disabled until an identified seller invitation flow exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contracts/{id}/status:
    patch:
      operationId: updateContractStatus
      tags: [contracts]
      summary: Legacy status update is disabled; state changes are server-verified
      parameters:
        - $ref: "#/components/parameters/ContractId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "409":
          description: Status cannot be changed directly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contracts/{id}/cancel:
    post:
      operationId: cancelUnfundedContract
      tags: [contracts]
      summary: Cancel an unfunded contract before any funding payment is active
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Unfunded contract cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContractCancellation" }
        "409":
          description: Funding is active or the contract is no longer cancellable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /contracts/{id}/fund:
    post:
      operationId: fundContract
      tags: [contracts]
      summary: Fund a Testnet escrow from the internal wallet balance
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Internal Testnet balance was debited and escrow funded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "409":
          description: Funding was blocked by contract state or wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Contract was not found for the authenticated buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /contracts/{id}/disputes:
    get:
      operationId: listDisputes
      tags: [disputes]
      summary: List disputes for a contract
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Disputes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dispute"
    post:
      operationId: createDispute
      tags: [disputes]
      summary: Open a dispute
      parameters:
        - $ref: "#/components/parameters/ContractId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Created dispute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [activity]
      summary: Mark one notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
        "404":
          description: Notification not found for the authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    piAppSessionCookie:
      type: apiKey
      in: cookie
      name: pi_app_session
  parameters:
    ContractId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ContractCancellation:
      type: object
      properties:
        cancelled: { type: boolean, const: true }
        contractId: { type: string, format: uuid }
        status: { type: string, const: cancelled }
      required: [cancelled, contractId, status]
...
    Contract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buyerId:
          type: ["string", "null"]
        sellerId:
          type: ["string", "null"]
        pipeline:
          type: string
          enum: [digital, shippable, local_property, custom_terms]
        metadata:
          type: object
          additionalProperties: true
        title:
          type: string
        reference:
          type: string
        buyerName:
          type: string
        sellerName:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, awaiting_funding, funded, submitted, in_delivery, completed, disputed, resolved, refunded, cancelled]
        dueDate:
          type: string
        paymentMethod:
          type: string
        nextAction:
          type: ["string", "null"]
        disputeCount:
          type: integer
        releaseDate:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - buyerId
        - sellerId
        - pipeline
        - metadata
        - title
        - reference
        - buyerName
        - sellerName
        - amount
        - currency
        - status
        - dueDate
        - paymentMethod
        - nextAction
        - disputeCount
        - releaseDate
        - submittedAt
        - createdAt
        - updatedAt
    ContractInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        reference:
          type: string
          minLength: 1
        buyerName:
          type: string
          minLength: 1
        sellerName:
          type: string
          minLength: 1
        amount:
          type: number
          exclusiveMinimum: 0
        currency:
          type: string
          pattern: "^(?:[Pp][Ii]|[A-Za-z]{3})$"
        dueDate:
          type: string
        paymentMethod:
          type: string
          minLength: 1
        pipeline:
          type: string
          enum: [digital, shippable, local_property, custom_terms]
        metadata:
          type: object
          additionalProperties: true
      required: [title, reference, buyerName, sellerName, amount, currency, dueDate, paymentMethod]
    StatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, awaiting_funding, funded, in_delivery, completed, disputed, resolved, refunded, cancelled]
        note:
          type: ["string", "null"]
      required: [status]
    Dispute:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        reason:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, resolved]
        requestedResolution:
          type: string
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, contractId, reason, description, status, requestedResolution, resolution, createdAt, updatedAt]
    DisputeInput:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        requestedResolution:
          type: string
          minLength: 1
      required: [reason, description, requestedResolution]
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        contractId: { type: string, format: uuid }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
        readAt: { type: ["string", "null"], format: date-time }
      required: [id, contractId, type, title, message, createdAt, readAt]
...